refactor(map-next): extract hash, basemap, and radius helpers
Pulls three self-contained, stable concerns out of the ~1000-line
+page.svelte into focused modules. Pure logic with no component-state
deps — chosen as the low-risk extractions ahead of the Phase 6.5
search/list work, which will add unrelated glue to +page.svelte.

- src/lib/map/mapHash.ts — HashCoords type, parseHashCoords,
  writeHashCoords. URL-hash viewport sync, window-only side effects.
- src/lib/map/basemaps.ts — BasemapId, BASEMAPS catalog,
  BASEMAP_STORAGE_KEY, isBasemapId, getStoredBasemap. The
  handleBasemapChange handler stays in the component (it touches
  the live map instance).
- viewport.ts gains calculateRadiusKmFromLngLatBounds — the
  MapLibre-shaped twin of the existing Leaflet calculateRadiusKm.
  The inlined copy in +page.svelte is removed.

+page.svelte drops ~100 lines. No behavior change; all gates green.
